% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_paths.R
\name{add_paths}
\alias{add_paths}
\title{Build Paths}
\usage{
add_paths(gps, truth, route)
}
\arguments{
\item{gps}{a gps file processed form GPX with \code{gpx_to_points}}

\item{truth}{a set on known locations for each point in route}

\item{route}{a vector of destinations}
}
\value{
gps object with added path ID
}
\description{
Add a new attribute to a GPX file processed with
}
