% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-drive-names.R
\name{fetch_second_drive}
\alias{fetch_second_drive}
\title{Fetch the name of the second HD as a string}
\usage{
fetch_second_drive()
}
\value{
Name as a string in the form "x:"
}
\description{
Fetch the name of the second HD as a string
}
