% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_same_paths.R
\name{find_same_paths}
\alias{find_same_paths}
\title{Find same paths within a vector of destinations}
\usage{
find_same_paths(route)
}
\arguments{
\item{route}{a vector of destinations}
}
\value{
a data.frame containing the path start and end nodes and order in route.
}
\description{
Find same paths within a vector of destinations
}
