% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_to_points.R
\name{gpx_to_pts}
\alias{gpx_to_pts}
\title{GPX file to sf object}
\usage{
gpx_to_pts(path)
}
\arguments{
\item{path}{a file path to a gpx file}
}
\value{
an sf points object with all gpx data attchaed
}
\description{
Given a GPX file return an sf points object
}
