% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h15 paths.R
\name{h15_paths}
\alias{h15_paths}
\title{Saves all required file paths to parent envrioment}
\usage{
h15_paths()
}
\value{
Multiple paths as string objects
}
\description{
Uses the ysi \code{paths} library and functions to write file paths strings to parent envrioment.
}
