% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mframe.R
\name{mframe}
\alias{mframe}
\alias{mframe.default}
\alias{mframe.lm}
\alias{mframe.glm}
\alias{mframe.pbart}
\alias{mframe.wbart}
\title{Extracting Model Frame}
\usage{
mframe(object, data, ...)

\method{mframe}{default}(object, data)

\method{mframe}{lm}(object, data = NULL)

\method{mframe}{glm}(object, data = NULL)

\method{mframe}{pbart}(object, data)

\method{mframe}{wbart}(object, data)
}
\arguments{
\item{object}{fitted model object}

\item{data}{data set containing all variables}
}
\value{
model frame without the outcome variable.
}
\description{
Extracing model frame from lm, glm, pbart, and wbart objects.
}
