% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{summary.paths}
\alias{summary.paths}
\alias{print.summary.paths}
\title{Summarizing Output from Causal Paths Analysis}
\usage{
\method{summary}{paths}(object, ...)

\method{print}{summary.paths}(x, ...)
}
\description{
Summarizing Output from Causal Paths Analysis
}
