% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-package.R
\docType{package}
\name{paths-package}
\alias{paths-package}
\title{paths: An Imputation Approach to Estimating Path-specific Causal Effects}
\description{
This package implements an imputation approach to estimating path-specific causal
effects as detailed in Zhou and Yamamoto (2020). Statistical inference is conducted using the nonparametric
bootstrap.
}
\references{
Zhou, Xiang and Teppei Yamamoto. 2020. "\href{https://osf.io/2rx6p}{Tracing Causal Paths from Experimental and Observational Data}".
}
\author{
Minh Trinh, Massachusetts Institute of Technology,
  \email{mdtrinh@mit.edu};
  Teppei Yamamoto, Massachusetts Institute of Technology,
  \email{teppei@mit.edu};
  Xiang Zhou, Harvard University,
  \email{xiang_zhou@fas.harvard.edu}
}
