% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-package.R
\docType{package}
\name{paths-package}
\alias{paths-package}
\title{paths: A package for causal paths analysis.}
\description{
The \code{paths} package provides a framework for
tracing causal paths in the presence of multiple
causally independent mediators. It implements the
identification and estimation strategies outlined
in Xiang & Yamamoto (2020), and conducts statistical
inference using non-parametric bootstrap.
}
\author{
Teppei Yamamoto, Massachusetts Institute of Technology,
  \email{teppei@mit.edu};
  Xiang Zhou, Harvard University,
  \email{xiang_zhou@fas.harvard.edu};
  Minh Trinh, Massachusetts Institute of Technology,
  \email{mdtrinh@mit.edu}
}
