% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{plot.paths}
\alias{plot.paths}
\title{Plot Method for \code{paths} Objects}
\usage{
\method{plot}{paths}(x, mediator_names = NULL, type = c(1, 2))
}
\arguments{
\item{x}{object of class \code{paths} as produced by the
\code{paths} function}

\item{mediator_names}{a vector of strings indicating the
labels for each mediator. Must contain as many elements
as there are mediators in the model. If not supplied,
the default labels will be constructed from the model
formulas supplied to the \code{paths} function.}

\item{type}{either \code{1}, \code{2}, or \code{c(1,2)},
indicating whether the plot will display estimates obtained
using Type I, Type II, or both Type I and Type II decompsitions.
Default is to show estimates from both types.}
}
\description{
Plot point estimates and confidence intervals
for each individual path-specific effect
from a \code{paths} object.
}
