% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.paths.R
\name{plot.paths}
\alias{plot.paths}
\title{#####################################################
Plot method for paths#' objects
#####################################################}
\usage{
\method{plot}{paths}(
  paths_obj,
  mediator_names = NULL,
  type = c(1, 2),
  estimator = c("pure", "hybrid"),
  horizontal = FALSE
)
}
\arguments{
\item{paths_obj}{object of class \code{paths} as produced by the \code{paths}
function}

\item{mediator_names}{a vector of strings indicating the labels for each
mediator. Must contain as many elements as there are mediators in the
model. If not supplied, the default labels will be constructed from the
model formulas supplied to the \code{paths} function.}

\item{type}{either \code{1}, \code{2}, or \code{c(1,2)}, indicating whether
the plot will display estimates obtained using Type I, Type II, or both
Type I and Type II decompsition. Default is to show estimates from both
types of decomposition.}

\item{estimator}{either \code{"pure"}, \code{"hybrid"}, or \code{c("pure",
"hybrid")}, indicating whether the plot will display estimates obtained
using the pure imputation estimator, the imputation-based weighting
estimator, or both. Default is to show estimates from both estimators.}

\item{horizontal}{logical indicating whether a horizontal plot should be
generated.}
}
\value{
a \code{ggplot2} plot, which can be further customized by the user.
}
\description{
Plot Method for \code{paths} Objects
}
\details{
Plot point estimates and confidence intervals for each individual
path-specific effect from a \code{paths} object.
}
