% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\alias{pred.lm}
\alias{pred.glm}
\alias{pred.pbart}
\alias{pred.wbart}
\title{Getting predicted values}
\usage{
pred(object, newdata, ...)

\method{pred}{lm}(object, newdata, ...)

\method{pred}{glm}(object, newdata, ...)

\method{pred}{pbart}(object, newdata, ...)

\method{pred}{wbart}(object, newdata, ...)
}
\arguments{
\item{object}{fitted model object}

\item{newdata}{a data frame containing predictors}
}
\value{
a vector of predicted values for \code{newdata}
}
\description{
Getting predicted values from lm, glm, pbart, and wbart objects.
}
