% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{summary.paths}
\alias{summary.paths}
\alias{print.summary.paths}
\title{Summarizing Output from Causal Paths Analysis}
\usage{
\method{summary}{paths}(object, ...)

\method{print}{summary.paths}(x, ...)
}
\arguments{
\item{object}{an object of class \code{paths} as generated by the
\code{paths} function.}

\item{...}{additional arguments (not yet implemented)}

\item{x}{an object of class \code{summary.paths}}
}
\value{
An object of class \code{summary.paths}, which is a list containing
  the \code{call}, \code{treat}, \code{outcome}, \code{mediators},
  \code{covariates}, \code{formulas}, \code{conditional}, \code{ps},
  \code{ps_formula}, \code{sims}, \code{conf.level} components from the
  \code{paths} object, plus \describe{ \item{nobs}{the number of
  observations.} \item{estimates}{a list containing two matrices,
  \code{estimates_t1} and \code{estimates_t2}, corresponding to effect
  estimates obtained using Type I and Type II decomposition. Each matrix
  contains the point estimates, confidence intervals and p-values of the
  total effect, direct effect, and each individual indirect effect
  corresponding to the appropriate decomposition. The elements in each matrix
  are extracted from the\code{est}, \code{CIs} and \code{p} components from
  the \code{paths} object. } }
}
\description{
Function to report results from causal paths analysis. Report point estimates
and standard errors for the total effect, direct effect, and each individual
indirect effect, separately for Type I and Type II decomposition.
}
\details{
\code{print.summary.paths} tries to smartly format the point
  estimates and confidence intervals, and provides 'signficance stars'
  through the \code{stats::printCoefmat} function.

  It also prints out the names of the outcome, treatment, mediator and
  covariate variables as extracted from the \code{formulas} argument of the
  call to \code{paths} for users to verify if the model formulas have been
  correctly specified.
}
\seealso{
\code{\link{paths}}, \code{\link{plot.paths}},
  \code{\link{summary}}.
}
\author{
Minh Trinh, Massachusetts Institute of Technology,
  \email{mdtrinh@mit.edu}
}
