% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tatar}
\alias{tatar}
\title{tatar dataset}
\format{An object of class \code{data.frame} with 427 rows and 22 columns.}
\usage{
tatar
}
\description{
A dataset containing the tatar data
}
\keyword{datasets}
