% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{welfare}
\alias{welfare}
\title{welfare dataset}
\format{An object of class \code{data.frame} with 213 rows and 26 columns.}
\usage{
welfare
}
\description{
A dataset containing the welfare data
}
\keyword{datasets}
